fix repo region move test folder
